requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pypdf==4.0.1
langchain>=0.1.0,<0.3
langchain-groq>=0.0.3
langchain-google-genai>=0.0.6
sentence-transformers==2.3.1
chromadb==0.4.22
streamlit==1.31.0
jira==3.6.0
PyGithub==2.1.1
pydantic==2.5.3
python-dotenv==1.0.0
json-repair>=0.7.0
pytest>=7.4.0
fpdf2>=2.7.0
